requirements.txt
spacy==3.8.5
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
matplotlib==3.10.1
pandas==2.2.3
scikit-learn==1.6.1
python-docx==1.1.2
PyPDF==3.0.1
joblib==1.3.2
streamlit==1.45.0
